requirements.txt
langchain
langchain_community
langchain-openai
faiss-cpu
pypdf
python-dotenv
streamlit
icecream